+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * Update User Operator to use Admin API for managing SCRAM-SHA-512 users
 * Configure fixed size limit for `emptyDir` volumes used for temporary files (#5340)
 * Update Strimzi Kafka Bridge to 0.20.2
+* Expose JMX port on Zookeeper nodes via a headless service
 
 ### Changes, deprecations and removals